+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Version - 0]
 
+## Features [v0.9.5] - Release Date: 2023-10-31
+
+- fix delete & update highlight news
+- clear mock data
+
 ## Features [v0.9.4] - Release Date: 2023-10-31
 
 - fix highlight news